requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
docopt==0.6.2
hdfs==2.6.0
idna==3.3
numpy==1.22.2
pandas==1.4.1
psycopg2-binary==2.9.3
py4j==0.10.9.3
pyspa==1.4
pyspark==3.2.1
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
six==1.16.0
urllib3==1.26.8
